refactor(BridgeMaker): BridgeMaker 클래스 디렉터리 이동
util -> bridge 기본 디렉터리
